+++ components/color-picker-field/index.scss
